Refactor WebSocket proxy logic to shared internal utilities
