+++ src/prefect/logs/clients.py
@@ -1,12 +1,9 @@
 import asyncio
-import os
 import ssl
 from datetime import timedelta
 from types import TracebackType
 from typing import (
     TYPE_CHECKING,
-    Any,
-    Generator,
     MutableMapping,
     Optional,
     Tuple,
@@ -14,17 +11,15 @@ from typing import (
     cast,
 )
 from urllib.parse import urlparse
-from urllib.request import proxy_bypass
 from uuid import UUID
 
 import certifi
 import orjson
 from cachetools import TTLCache
 from prometheus_client import Counter
-from python_socks.async_.asyncio import Proxy
 from typing_extensions import Self
 from websockets import Subprotocol
-from websockets.asyncio.client import ClientConnection, connect
+from websockets.asyncio.client import ClientConnection
 from websockets.exceptions import (
     ConnectionClosed,
     ConnectionClosedError,
@@ -32,6 +27,7 @@ from websockets.exceptions import (
 )
 
 import prefect.types._datetime
+from prefect._internal.websockets import websocket_connect
 from prefect.client.schemas.objects import Log
 from prefect.logging import get_logger
 from prefect.settings import (
@@ -77,70 +73,25 @@ def logs_out_socket_from_api_url(url: str) -> str:
     return http_to_ws(url) + "/logs/out"
 
 
-class WebsocketProxyConnect(connect):
-    def __init__(self: Self, uri: str, **kwargs: Any):
-        # super() is intentionally deferred to the _proxy_connect method
-        # to allow for the socket to be established first
+def _create_ssl_context_for_websocket(uri: str) -> ssl.SSLContext | None:
+    """Create SSL context for WebSocket connections based on URI scheme."""
+    u = urlparse(uri)
 
-        self.uri = uri
-        self._kwargs = kwargs
+    if u.scheme != "wss":
+        return None
 
-        u = urlparse(uri)
-        host = u.hostname
-
-        if not host:
-            raise ValueError(f"Invalid URI {uri}, no hostname found")
-
-        if u.scheme == "ws":
-            port = u.port or 80
-            proxy_url = os.environ.get("HTTP_PROXY")
-        elif u.scheme == "wss":
-            port = u.port or 443
-            proxy_url = os.environ.get("HTTPS_PROXY")
-            kwargs["server_hostname"] = host
-        else:
-            raise ValueError(
-                "Unsupported scheme %s. Expected 'ws' or 'wss'. " % u.scheme
-            )
-
-        self._proxy = (
-            Proxy.from_url(proxy_url) if proxy_url and not proxy_bypass(host) else None
-        )
-        self._host = host
-        self._port = port
-
-        if PREFECT_API_TLS_INSECURE_SKIP_VERIFY and u.scheme == "wss":
-            # Create an unverified context for insecure connections
-            ctx = ssl.create_default_context()
-            ctx.check_hostname = False
-            ctx.verify_mode = ssl.CERT_NONE
-            self._kwargs.setdefault("ssl", ctx)
-        elif u.scheme == "wss":
-            cert_file = PREFECT_API_SSL_CERT_FILE.value()
-            if not cert_file:
-                cert_file = certifi.where()
-            # Create a verified context with the certificate file
-            ctx = ssl.create_default_context(cafile=cert_file)
-            self._kwargs.setdefault("ssl", ctx)
-
-    async def _proxy_connect(self: Self) -> ClientConnection:
-        if self._proxy:
-            sock = await self._proxy.connect(
-                dest_host=self._host,
-                dest_port=self._port,
-            )
-            self._kwargs["sock"] = sock
-
-        super().__init__(self.uri, **self._kwargs)
-        proto = await self.__await_impl__()
-        return proto
-
-    def __await__(self: Self) -> Generator[Any, None, ClientConnection]:
-        return self._proxy_connect().__await__()
-
-
-def websocket_connect(uri: str, **kwargs: Any) -> WebsocketProxyConnect:
-    return WebsocketProxyConnect(uri, **kwargs)
+    if PREFECT_API_TLS_INSECURE_SKIP_VERIFY.value():
+        # Create an unverified context for insecure connections
+        ctx = ssl.create_default_context()
+        ctx.check_hostname = False
+        ctx.verify_mode = ssl.CERT_NONE
+        return ctx
+    else:
+        # Create a verified context with the certificate file
+        cert_file = PREFECT_API_SSL_CERT_FILE.value()
+        if not cert_file:
+            cert_file = certifi.where()
+        return ssl.create_default_context(cafile=cert_file)
 
 
 def _get_api_url_and_key(
@@ -248,10 +199,13 @@ class PrefectLogsSubscriber:
 
         logger.debug("Connecting to %s", socket_url)
 
-        self._connect = websocket_connect(
-            socket_url,
-            subprotocols=[Subprotocol("prefect")],
-        )
+        # Configure SSL context for the connection
+        ssl_context = _create_ssl_context_for_websocket(socket_url)
+        connect_kwargs = {"subprotocols": [Subprotocol("prefect")]}
+        if ssl_context:
+            connect_kwargs["ssl"] = ssl_context
+
+        self._connect = websocket_connect(socket_url, **connect_kwargs)
         self._websocket = None
         self._reconnection_attempts = reconnection_attempts
         if self._reconnection_attempts < 0:
